Slightly rephrase other results link message
Ew duplicate code. Should probably have a method in the SciolyFF
interpreter for these kinds of names, but oh well for now.

Is this exact phrasing important? Not really, except that because this
exact phrasing is not important, it probably won't ever be important
enough to change, so what I choose now will probably be the phrasing
f o r e v e r so yes thank you it is very important indeed.